MID-2836: Route Weights
  * Document Weighting of routes
  * Update all generated routes to have large drivide weights associated with them
  * Add tests to check for precedence in the case ofdynamic predicates added in Skipper.
