+++ src/test/scala/ie/zalando/fabric/gateway/service/IngressDerivationChainSpec.scala
@@ -1,11 +1,12 @@
 package ie.zalando.fabric.gateway.service
 
 import akka.actor.ActorSystem
-import akka.http.scaladsl.model.Uri
 import akka.stream.ActorMaterializer
 import cats.data.{NonEmptyList => NEL}
 import ie.zalando.fabric.gateway.models.SynchDomain._
+import ie.zalando.fabric.gateway.service.RouteDerivationModels.Admin
 import ie.zalando.fabric.gateway.web.marshalling.JsonModels
+import ie.zalando.fabric.gateway.service.TestUtils._
 import org.mockito.scalatest.MockitoSugar
 import org.scalatest.prop.TableDrivenPropertyChecks._
 import org.scalatest.{FlatSpec, Matchers}
@@ -23,168 +24,6 @@ class IngressDerivationChainSpec extends FlatSpec with MockitoSugar with Matcher
   val kubernetesClient       = mock[KubernetesClient]
   val stackSetOperations     = new StackSetOperations(kubernetesClient)
   val ingressDerivationLogic = new IngressDerivationChain(stackSetOperations, None)
-
-  val AdminUser                 = "adminUser"
-  val WhitelistedUser           = "whitelistedUser"
-  val ResourceWhitelistedUser   = "resourceWhitelistedUser"
-  val InheritedWhitelistDetails = WhitelistConfig(Set(), Inherited)
-  val UserWhitelist             = EmployeeAccessConfig(Set.empty)
-  val EnabledCors = Some(
-    CorsConfig(Set(Uri.from(host = "example.com"), Uri.from(host = "example-other.com")),
-               Set("Content-Type", "Authorization", "X-Flow-id")))
-  val DisabledCors: Option[CorsConfig] = None
-
-  val sampleGateway = GatewaySpec(
-    SchemaDefinedServices(Set(IngressBackend("host", Set(ServiceDescription("svc"))))),
-    Set(AdminUser),
-    WhitelistConfig(Set(), Disabled),
-    DisabledCors,
-    Map(
-      PathMatch("/api/resource") -> PathConfig(
-        Map(
-          Get -> ActionAuthorizations(
-            NEL.of("uid", "service.read"),
-            Some(RateLimitDetails(10, PerMinute, Map.empty[String, Int])),
-            InheritedWhitelistDetails,
-            UserWhitelist
-          ),
-          Post -> ActionAuthorizations(
-            NEL.of("uid", "service.write"),
-            Some(
-              RateLimitDetails(10,
-                               PerMinute,
-                               Map(
-                                 AdminUser   -> 25,
-                                 "otherUser" -> 35
-                               ))),
-            InheritedWhitelistDetails,
-            UserWhitelist
-          )
-        )),
-      PathMatch("/api/resource/*") -> PathConfig(
-        Map(
-          Get -> ActionAuthorizations(
-            NEL.of("uid", "service.read"),
-            None,
-            InheritedWhitelistDetails,
-            UserWhitelist
-          )
-        ))
-    )
-  )
-
-  val sampleGloballyWhitelistedGateway = GatewaySpec(
-    SchemaDefinedServices(Set(IngressBackend("host", Set(ServiceDescription("svc", NamedServicePort("named")))))),
-    Set(AdminUser),
-    WhitelistConfig(Set(WhitelistedUser), Enabled),
-    DisabledCors,
-    Map(
-      PathMatch("/api/resource") -> PathConfig(
-        Map(
-          Get -> ActionAuthorizations(
-            NEL.of("uid", "service.read"),
-            Some(RateLimitDetails(10, PerMinute, Map.empty[String, Int])),
-            InheritedWhitelistDetails,
-            UserWhitelist
-          ),
-          Post -> ActionAuthorizations(
-            NEL.of("uid", "service.write"),
-            Some(
-              RateLimitDetails(10,
-                               PerMinute,
-                               Map(
-                                 AdminUser       -> 25,
-                                 WhitelistedUser -> 25,
-                                 "otherUser"     -> 35
-                               ))),
-            InheritedWhitelistDetails,
-            UserWhitelist
-          )
-        )),
-      PathMatch("/api/resource/*") -> PathConfig(
-        Map(
-          Get -> ActionAuthorizations(
-            NEL.of("uid", "service.read"),
-            None,
-            InheritedWhitelistDetails,
-            UserWhitelist
-          ))
-      )
-    )
-  )
-
-  val sampleResourceWhitelistingGateway = GatewaySpec(
-    SchemaDefinedServices(Set(IngressBackend("host", Set(ServiceDescription("svc", NamedServicePort("named")))))),
-    Set(AdminUser),
-    WhitelistConfig(Set(WhitelistedUser), Enabled),
-    DisabledCors,
-    Map(
-      PathMatch("/api/resource") -> PathConfig(
-        Map(
-          Get -> ActionAuthorizations(
-            NEL.of("uid", "service.read"),
-            Some(RateLimitDetails(10, PerMinute, Map.empty[String, Int])),
-            InheritedWhitelistDetails,
-            UserWhitelist
-          ),
-          Post -> ActionAuthorizations(
-            NEL.of("uid", "service.write"),
-            Some(
-              RateLimitDetails(10,
-                               PerMinute,
-                               Map(
-                                 AdminUser       -> 25,
-                                 WhitelistedUser -> 25,
-                                 "otherUser"     -> 35
-                               ))),
-            WhitelistConfig(Set(ResourceWhitelistedUser), Enabled),
-            UserWhitelist
-          )
-        )),
-      PathMatch("/api/resource/*") -> PathConfig(
-        Map(
-          Get -> ActionAuthorizations(
-            NEL.of("uid", "service.read"),
-            None,
-            WhitelistConfig(Set(), Disabled),
-            UserWhitelist
-          )
-        ))
-    )
-  )
-
-  val sampleUserWhitelistingGateway = GatewaySpec(
-    SchemaDefinedServices(Set(IngressBackend("host", Set(ServiceDescription("svc", NamedServicePort("named")))))),
-    Set(AdminUser),
-    WhitelistConfig(Set(WhitelistedUser), Enabled),
-    DisabledCors,
-    Map(
-      PathMatch("/api/resource") -> PathConfig(
-        Map(
-          Get -> ActionAuthorizations(
-            NEL.of("uid", "service.read"),
-            None,
-            InheritedWhitelistDetails,
-            EmployeeAccessConfig(Set(WhitelistedUser))
-          ),
-          Post -> ActionAuthorizations(
-            NEL.of("uid", "service.write"),
-            None,
-            WhitelistConfig(Set(ResourceWhitelistedUser), Enabled),
-            UserWhitelist
-          )
-        )),
-      PathMatch("/api/resource/*") -> PathConfig(
-        Map(
-          Get -> ActionAuthorizations(
-            NEL.of("uid", "service.read"),
-            Some(RateLimitDetails(10, PerMinute, Map.empty[String, Int])),
-            WhitelistConfig(Set(), Disabled),
-            EmployeeAccessConfig(Set(WhitelistedUser))
-          )
-        ))
-    )
-  )
 
   val testableRouteDerivationWithCatchAll: List[IngressDefinition] =
     Await.result(ingressDerivationLogic.deriveRoutesFor(sampleGateway,
@@ -260,23 +99,24 @@ class IngressDerivationChainSpec extends FlatSpec with MockitoSugar with Matcher
     routes.size shouldBe 3
 
     routes.foreach { sr =>
-      sr.predicates.size should be(4)
+      sr.predicates.size should be(5)
       sr.predicates.exists { _.getClass == classOf[PathMatch] } should be(true)
       sr.predicates.exists { _.getClass == classOf[MethodMatch] } should be(true)
       sr.predicates.exists { _.getClass == classOf[UidMatch] } should be(true)
+      sr.predicates.exists { _.getClass == classOf[WeightedRoute] } should be(true)
       sr.filters should equal(
         List(NonCustomerRealm, EnableAccessLog(List(2, 4, 5)), AccessLogAuditing(), RequiredPrivileges(NEL.of("uid")), FlowId, ForwardTokenInfo))
     }
 
     routes
       .map(_.predicates)
-      .contains(List(PathMatch("/api/resource"), MethodMatch(Get), UidMatch(NEL.one(AdminUser)), HttpsTraffic)) shouldBe true
+      .contains(List(PathMatch("/api/resource"), MethodMatch(Get), UidMatch(NEL.one(AdminUser)), HttpsTraffic, WeightedRoute(Admin))) shouldBe true
     routes
       .map(_.predicates)
-      .contains(List(PathMatch("/api/resource"), MethodMatch(Post), UidMatch(NEL.one(AdminUser)), HttpsTraffic)) shouldBe true
+      .contains(List(PathMatch("/api/resource"), MethodMatch(Post), UidMatch(NEL.one(AdminUser)), HttpsTraffic, WeightedRoute(Admin))) shouldBe true
     routes
       .map(_.predicates)
-      .contains(List(PathMatch("/api/resource/*"), MethodMatch(Get), UidMatch(NEL.one(AdminUser)), HttpsTraffic)) shouldBe true
+      .contains(List(PathMatch("/api/resource/*"), MethodMatch(Get), UidMatch(NEL.one(AdminUser)), HttpsTraffic, WeightedRoute(Admin))) shouldBe true
   }
 
   "Route Filtering" should "not generate rate limits for admins" in {
@@ -760,46 +600,5 @@ class IngressDerivationChainSpec extends FlatSpec with MockitoSugar with Matcher
     filteredRoutes.foreach { filters: NEL[SkipperFilter] =>
       filters.toList should contain(AccessLogAuditing("sub"))
     }
-  }
-
-  def isAdminRoute(route: IngressDefinition): Boolean = {
-    val defn = route.metadata.routeDefinition
-    defn.customRoute.isEmpty &&
-    !defn.filters.exists(_.getClass == classOf[GlobalRouteRateLimit]) &&
-    !defn.filters.exists(_.getClass == classOf[ClientSpecificRouteRateLimit]) &&
-    route.metadata.name.endsWith("admins")
-  }
-
-  def isCorsRoute(route: IngressDefinition): Boolean = {
-    val metadata = route.metadata
-    metadata.name.contains("-cors") && metadata.routeDefinition.customRoute.isDefined
-  }
-
-  def isWhitelistedUserRoute(route: IngressDefinition): Boolean = {
-    val defn = route.metadata.routeDefinition
-    defn.customRoute.isEmpty &&
-    !defn.filters.exists(_.getClass == classOf[GlobalRouteRateLimit]) &&
-    !defn.filters.exists(_.getClass == classOf[ClientSpecificRouteRateLimit]) &&
-    route.metadata.name.endsWith("-users-all")
-  }
-
-  def isCatchAllRoute(route: IngressDefinition): Boolean = {
-    route.metadata.routeDefinition.customRoute.exists { customRoute =>
-      NEL.of(Status(404), DefaultRejectMsg, Shunt).forall(customRoute.filters.toList.contains)
-    }
-  }
-
-  def isHttpRejectRoute(route: IngressDefinition): Boolean = {
-    route.metadata.routeDefinition.customRoute.exists { customRoute =>
-      NEL.of(Status(400), HttpRejectMsg, Shunt).forall(customRoute.filters.toList.contains)
-    }
-  }
-
-  def isServiceSpecificRoute(route: IngressDefinition): Boolean = {
-    route.metadata.routeDefinition.predicates.exists(_.getClass == classOf[ClientMatch])
-  }
-
-  def isWhitelistRejectRoute(route: IngressDefinition): Boolean = {
-    route.metadata.name.endsWith("-non-whitelisted")
   }
 }
